requirements.txt
aiofiles==23.1.0
aiohttp==3.7.3
albumentations==0.5.2
altair==5.0.1
appdirs==1.4.4
argcomplete==3.0.8
atomicwrites==1.4.0
attrs==20.3.0
audioread==3.0.0
backoff==1.10.0
bkcharts==0.2
black==19.10b0
blessed==1.20.0
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.2.0
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
clyent==1.2.2
cmake==3.25.0
configparser==5.0.1
constantly==15.1.0
croniter==1.3.15
cssselect==1.1.0
cycler==0.10.0
cytoolz==0.11.0
daal4py==2021.5.0
dacite==1.7.0
datashape==0.5.4
dateutils==0.6.12
decorator==4.4.2
deepdiff==6.3.0
Deprecated==1.2.14
dill==0.3.6
dnspython==2.3.0
docopt==0.6.2
dominate==2.7.0
dotmap==1.3.30
einops==0.6.1
eventlet==0.33.3
fastapi==0.88.0
ffmpy==0.3.0
fiftyone==0.21.0
filelock==3.0.12
fire==0.5.0
fonttools==4.25.0
fsspec==2023.5.0
ftfy==6.1.1
gitdb==4.0.5
GitPython==3.1.12
gradio==1.6.4
gradio_client==0.2.5
grpcio==1.34.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==2.10
imgaug==0.4.0
inflate64==0.3.1
inflection==0.5.1
inquirer==3.1.3
ipdb==0.13.13
Jinja2==3.0.0
joblib==1.0.0
jsonlines==3.1.0
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
kaggle==1.5.10
kaleido==0.2.1
kiwisolver==1.3.1
lazy_loader==0.2
librosa==0.10.0.post2
lit==15.0.7
llvmlite==0.38.0
Markdown==3.3.3
markdown2==2.4.8
matplotlib==3.4.3
mccabe==0.6.1
mdurl==0.1.2
medmnist==2.1.0
mongoengine==0.24.2
monotonic==1.6
motor==3.1.2
mpmath==1.2.1
multidict==5.1.0
multivolumefile==0.2.3
munch==2.5.0
munkres==1.1.4
networkx==2.5
nltk==3.5
numpy==1.21.6
oauthlib==3.1.0
omegaconf==2.0.5
orjson==3.9.0
packaging==23.1
pandas==1.2.0
paramiko==3.2.0
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.2
pep8==1.7.1
Pillow==10.0.0
pooch==1.6.0
pprintpp==0.4.0
pretrainedmodels==0.7.4
priority==2.0.0
progressbar==2.5
promise==2.3
protobuf==3.14.0
py7zr==0.20.5
pyarrow==12.0.0
pybcj==1.0.1
pycocotools==2.0.6
pycosat==0.6.3
pycryptodomex==3.18.0
pycurl==7.44.1
pydantic==1.10.8
pydicom==2.1.2
PyDispatcher==2.0.5
pydub==0.25.1
Pygments==2.15.1
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==2.4.7
pypinyin==0.47.1
pyppmd==1.0.0
pysam==0.20.0
pytest==7.1.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==6.0
pyzstd==0.15.7
queuelib==1.5.0
rarfile==4.0
readchar==4.0.5
regex==2020.11.13
reportlab==4.0.4
requests==2.25.1
retrying==1.3.4
rich==13.4.1
rpm==0.0.2
rsa==4.7
sacred==0.8.4
sacremoses==0.0.43
scipy==1.6.0
setproctitle==1.3.2
Shapely==1.7.1
shortuuid==1.0.1
SimpleITK==2.2.0
sip==4.19.13
six==1.15.0
smmap==3.0.4
soundfile==0.12.1
soxr==0.3.5
starlette==0.22.0
starsessions==1.3.0
subprocess32==3.5.4
tabulate==0.8.9
TBB==0.2
tensorboard==2.4.1
termcolor==2.2.0
texttable==1.6.7
threadpoolctl==2.1.0
tifffile==2021.1.14
timm==0.4.12
tokenizers==0.9.4
torch==1.8.0+cu111
torchaudio==0.8.0
torchfile==0.1.0
torchvision==0.9.0+cu111
tqdm==4.65.0
traitlets==5.9.0
transformers==4.2.1
triton==2.0.0
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==1.26.2
uvicorn==0.22.0
visdom==0.1.8.9
wandb==0.10.14
watchdog==0.10.4
webencodings==0.5.1
websockets==11.0.3
Werkzeug==1.0.1
wsproto==1.2.0
xmltodict==0.13.0
zict==2.0.0